content-security-policy/script-src/script-src-1_4.html

<!DOCTYPE HTML>
<html>
<head>
    <script src='/resources/testharness.js'></script>
    <script src='/resources/testharnessreport.js'></script>
    <link rel='stylesheet' href='/resources/testharness.css'>
</head>
<body>
    <h1>eval() should not run without 'unsafe-eval' script-src directive.</h1>
    <div id='log'></div>

   	<script>

        var evalRan = false;

        try {
          eval('evalRan = true;');
        } catch(e) {
          console.log(e);   
        }

        test(function() {assert_false(evalRan, "Unsafe eval ran.")}, "eval() should be blocked without an 'unsafe-eval' directive");
    </script>

    <script async defer src='../support/checkReport.sub.js?reportField=violated-directive&reportValue=script-src%20%27self%27+%27unsafe-inline%27'></script>

</body>
</html>
